testslide remove from the home page
