+++ src/Pages/Home/Home.jsx
@@ -5,7 +5,6 @@ import Hero from "../../Components/Home/Hero";
 import OurProcess from "../../Components/Home/OurProcess";
 import PopularCourse from "../../Components/Home/PopularCourse";
 import Review from "../../Components/Home/Review";
-import TestSlide from "../../Components/Home/testSlide";
 
 const Home = () => {
     return (
@@ -15,7 +14,6 @@ const Home = () => {
             <Categories></Categories>
             <PopularCourse></PopularCourse>
             <Review></Review>
-            <TestSlide></TestSlide>
             <OurProcess></OurProcess>
             <Faq></Faq>
         </>
